focus: Analysis of first and second focus of the viewer
First focus is defined here as the first non-white space, the viewer looks
at when Image is displayed to him/her.
Second focus is defined as the second non-white space, the viewer looks at,
whose color is different that the color of the first-focus space.

Data cleaned and extracted at focus/first-focus.csv and
focus/second-focus.csv and plots created are added at plots/focus/
directory.

Signed-off-by: Prathamesh Chavan <pc44800@gmail.com>