EyeTracker: Added file to the repository
EyeTracker.cpp was used with the help of Visual-Studio and
OpenCV, along with the eye-tracker to obtain the data
of where the user is looking. When executed, the user is
shown a set of 30 images picked from the set of 120 images,
depending upon the PERSON_ID variable. The viewer looks in the
images and we obtain the co-ordinates of where the user
is looking. This output data is stored in a separate .csv file,
for each image. Data stored in raw/ is collected in the
same manner.

Signed-off-by: Prathamesh Chavan <pc44800@gmail.com>